Finished auth, middleware
- Finished tying user notes to a given user
- Added lots of middleware
- Fixed some formatting

+++ backend/src/controllers/notesControllers.ts
@@ -2,10 +2,17 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+
+        
+
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -14,17 +21,23 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
 
-        if(!mongoose.isValidObjectId(noteId)) {
+        if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note ID");
         }
 
         const note = await NoteModel.findById(noteId).exec();
 
-        if(!note) {
+        if (!note) {
             throw createHttpError(404, "Note not found!");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
         }
 
         res.status(200).json(note);
@@ -39,17 +52,20 @@ interface CreateNoteBody {
 }
 
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
-
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try {
-        if(!title) {
+        assertIsDefined(authenticatedUserId);
+
+        if (!title) {
             throw createHttpError(400, "Note must have a title!");
         }
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
-            text: text, 
+            text: text,
         });
 
         res.status(201).json(newNote);
@@ -70,20 +86,26 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
 
-        if(!mongoose.isValidObjectId(noteId)) {
+        if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note ID");
         }
-        if(!newTitle) {
+        if (!newTitle) {
             throw createHttpError(400, "Note must have a title!");
         }
 
         const note = await NoteModel.findById(noteId).exec();
 
-        if(!note) {
+        if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
         }
 
         note.title = newTitle;
@@ -96,17 +118,24 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     }
 }
 
-export const deleteNote: RequestHandler = async(req, res, next) => {
+export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+    
 
     try {
-        if(!mongoose.isValidObjectId(noteId)) {
+        assertIsDefined(authenticatedUserId);
+        if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
         const note = await NoteModel.findById(noteId).exec();
 
-        if(!note) {
+        if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
         }
 
         await note.deleteOne();
